added ANSI/linux variant
fixed license text
