+++ LICENSE
@@ -1,5 +1,5 @@
 Copyright (c) 1983, KRVM
-Copyright (c) 2017, Johnathon Haves
+Copyright (c) 2016, Johnathon Havens
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
